Add customer ops ROIAlign and nms into Autocast (#9)

+++ torch_ipex/csrc/cpu/ExtendOPs.h
@@ -11,29 +11,29 @@ class AtenIpexTypeExt {
  public:
   static at::Tensor ROIAlign_forward(const at::Tensor& input,
                                      const at::Tensor& rois,
-                                     const float spatial_scale,
-                                     const int pooled_height,
-                                     const int pooled_width,
-                                     const int sampling_ratio);
+                                     const double spatial_scale,
+                                     const int64_t pooled_height,
+                                     const int64_t pooled_width,
+                                     const int64_t sampling_ratio);
 
   static at::Tensor ROIAlign_backward(const at::Tensor& grad,
                                       const at::Tensor& rois,
-                                      const float spatial_scale,
-                                      const int pooled_height,
-                                      const int pooled_width,
-                                      const int batch_size,
-                                      const int channels,
-                                      const int height,
-                                      const int width,
-                                      const int sampling_ratio);
+                                      const double spatial_scale,
+                                      const int64_t pooled_height,
+                                      const int64_t pooled_width,
+                                      const int64_t batch_size,
+                                      const int64_t channels,
+                                      const int64_t height,
+                                      const int64_t width,
+                                      const int64_t sampling_ratio);
     
   static at::Tensor nms(const at::Tensor& dets,
                         const at::Tensor& scores,
-                        const float threshold);
+                        const double threshold);
 
   static std::tuple<at::Tensor, at::Tensor, at::Tensor> batch_score_nms(const at::Tensor& dets,
                         const at::Tensor& scores,
-                        const float threshold);
+                        const double threshold);
 
   static at::Tensor interaction_forward(const std::vector<at::Tensor> & input);
   static std::vector<at::Tensor> interaction_backward(const at::Tensor & grad_out,
